First Commit
I’m at least happy with it to the point I’m okay with uploading my
progress. Detects presses, detects the time between presses and the
length of presses, and uses the ratios of button presses and dead air
to determine the letters. It prints e at the beginning because I made
it push upon loading to clear the deadair time accumulated due to
compiling and running.